fix(pg): use the new pgdg repo path
As of 15 April 2019, there is only one repository RPM per distro,
and it includes repository information for all available PostgreSQL releases.

So, the old pgdg repositories are no longer available.

By the way, we delete code and files related to the
old way we had to detect the latest repo release.